+++ livre/core/Cache/EmptyCacheObject.h
@@ -50,7 +50,7 @@ public:
     /**
      * @return The empty size.
      */
-    virtual uint32_t getCacheSize( ) const { return 0; }
+    virtual size_t getCacheSize( ) const { return 0; }
 
 private:
